Renaming theme directory due to conflict with existing wordpress theme.
There already exists a WP Spark theme on https://wordpress.org/themes/spark/ and loading this Spark theme kept suggesting there was an upgrade to the version of the theme, conflicting with the one already present on WordPress.com. So this changes the namespace for the theme directory to weerd.spark.